project/requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.9
conda==4.3.16
distlib==0.3.4
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-spectacular==0.21.1
filelock==3.4.2
idna==3.3
inflection==0.5.1
jsonschema==4.4.0
pipenv==2021.11.23
platformdirs==2.4.1
pycosat==0.6.3
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.11.0
virtualenv-clone==0.5.7
